fix: #293 ai summary
