+++ src/main/java/com/umc/linkyou/converter/AiArticleConverter.java
@@ -6,7 +6,7 @@ import com.umc.linkyou.domain.classification.Category;
 import com.umc.linkyou.domain.classification.Emotion;
 import com.umc.linkyou.domain.classification.Situation;
 import com.umc.linkyou.domain.mapping.UsersLinku;
-import com.umc.linkyou.infra.ai.dto.SummaryAnalysisResultDTO;
+import com.umc.linkyou.gemini.dto.SummaryResultDTO;
 import com.umc.linkyou.web.dto.AiArticleResponsetDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class AiArticleConverter {
     // 엔티티 생성 메서드
-    public static AiArticle toEntity(SummaryAnalysisResultDTO result,
+    public static AiArticle toEntity(SummaryResultDTO result,
                                      Situation selectedSituation,
                                      Emotion selectedEmotion,
                                      Category selectedCategory,
@@ -30,6 +30,7 @@ public class AiArticleConverter {
                 .aiCategoryId(selectedCategory.getCategoryId())
                 .summary(result.getSummary())
                 .imgUrl(imageUrl)
+                .keyword(result.getKeywords()) // [추가] 키워드 저장도 잊지 마세요!
                 .build();
     }
