+++ src/main/java/com/umc/linkyou/infra/ai/classifier/GeminiCategoryClassifier.java
@@ -2,8 +2,8 @@ package com.umc.linkyou.infra.ai.classifier;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.umc.linkyou.infra.ai.GeminiJsonUtils;
-import com.umc.linkyou.infra.parser.TitleDomainParser;
-import com.umc.linkyou.infra.parser.WebContentExtractor;
+import com.umc.linkyou.utils.parser.TitleDomainParser;
+import com.umc.linkyou.utils.parser.WebContentExtractor;
 import com.umc.linkyou.service.curation.gemini.GeminiTextService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
